requirements.txt
tensorflow>=2.0.0
opencv-python>=4.0.0
numpy>=1.19.0
pillow>=8.0.0
scikit-learn>=0.24.0
